+++ CHANGELOG.md
@@ -19,7 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Multiple files selection with image picker
 * Date input component
 * Switcher component
-* Radio group component 
+* `onScroll` and `onScrollBottom` props in chat component
+* Radio group component
 
 ### Changed